agents/supplier-agent/requirements.txt
fastapi
uvicorn
google-generativeai
psycopg2-binary
sqlalchemy
requests
python-dotenv
pydantic
langfuse
